Move enumerations to different package; Add accessCategoryToVisibleTo/AccessibleTo method
